+++ kaohe/A/A.h
@@ -1,5 +1,5 @@
 #ifndef _A_H
 #define _A_H
-int sum_A(int a, int b);
+double	sum_A(double a, double b);
 
 #endif
